requirements.txt
MySQL-python
pyyaml
